refactor: Move all types into types.ts file
