--- src/system/consts.ts
@@ -1,5 +1,0 @@
-const EVENTS = ['chat', 'sub', 'resub', 'cheer', 'subgift'] as const;
-
-const STATUS = ['idle', 'loading', 'running'] as const;
-
-export { EVENTS, STATUS };
