+++ src/system/store.ts
@@ -1,6 +1,5 @@
 import { create } from 'zustand';
-import { FieldName, ModeName } from './spy';
-import { STATUS } from './consts';
+import { FieldName, ModeName, Status } from './types';
 
 // Mode
 const useModeStore = create<ModeName>()(() => 'users');
@@ -13,7 +12,6 @@ const useInputStore = create<Record<FieldName, string[]>>()(() => ({
 }));
 
 // Status
-type Status = (typeof STATUS)[number];
 const useStatusStore = create<Status>()(() => 'idle');
 
 // Spying State
